Remove email interface
